Move manual auth tests to their own file
